ignore invalid records

+++ flink-runtime/src/main/java/org/apache/flink/runtime/taskmanager/Task.java
@@ -1428,13 +1428,15 @@ public class Task
 
         if (executionState == ExecutionState.RUNNING) {
             checkState(invokable instanceof CheckpointableTask, "invokable is not checkpointable");
-            // block consumption for rescaling
-            for (InputGate gate : inputGates) {
-                gate.setSuspended(true);
-            }
-            // flush outputs
-            for (ResultPartitionWriter partitionWriter : consumableNotifyingPartitionWriters) {
-                partitionWriter.flushAll();
+            if(checkpointOptions.isRescaling()) {
+                // block consumption for rescaling
+                for (InputGate gate : inputGates) {
+                    gate.setSuspended(true);
+                }
+                // flush outputs
+                for (ResultPartitionWriter partitionWriter : consumableNotifyingPartitionWriters) {
+                    partitionWriter.flushAll();
+                }
             }
             try {
                 ((CheckpointableTask) invokable)
